Moved animal submission status to animal submission
